fix: use SHA1 for CID instead of a numerical index
RFC 2111 requires that Content IDs (CID) be globally unique. We were not
complying with this restriction and were using CIDs that were only unique within
a message. Interestingly this caused no known problems except on iOS, presumably
because the obvious way to implement CIDs in a mail client is to look within the
current message first when resolving CID links. However, iOS' mail app maintains
a CID cache and looks within this before looking in the current message. The
result was that it would occasionally show the wrong image when it happened to
find something (from another message) in its cache that matched the CID it was
looking for.

As we're now using a hash value, even if there is a duplicated CID mail clients
should still show the correct image (excepting the unlikely case of a hash
collision).

Thanks to Alexander Harkness (@bearbin) who posted a similar resolution to this
issue using UUIDs (Github #7).

Closes Github #4 "unique Content-IDs"